Split off system related stuff from FileUtil into SystemUtil
